NP-988 Update translations and layout
