+++ src/pages/search/Search.tsx
@@ -1,10 +1,9 @@
 import React, { FC } from 'react';
-
+import { useTranslation } from 'react-i18next';
+import { Typography } from '@material-ui/core';
 import useSearchRegistrations from '../../utils/hooks/useSearchRegistrations';
 import SearchResults from './SearchResults';
 import ListSkeleton from '../../components/ListSkeleton';
-import { useTranslation } from 'react-i18next';
-import { Typography } from '@material-ui/core';
 
 interface SearchProps {
   searchTerm?: string;
@@ -18,11 +17,9 @@ const Search: FC<SearchProps> = ({ searchTerm, noHitsText }) => {
   return (
     <>
       {isLoadingSearch || !registrationsSearch ? (
-        <ListSkeleton arrayLength={5} minWidth={40} height={100} />
+        <ListSkeleton arrayLength={3} minWidth={40} height={100} />
       ) : registrationsSearch.hits.length > 0 ? (
         <SearchResults searchResult={registrationsSearch} searchTerm={searchTerm} />
-      ) : noHitsText ? (
-        noHitsText
       ) : (
         <Typography>{noHitsText ? noHitsText : t('no_hits')}</Typography>
       )}
