+++ src/translations/nb/common.json
@@ -50,6 +50,8 @@
   "save": "Lagre",
   "save_and_present": "Lagre og presenter",
   "search": "Søk",
+  "search_summary": "{{count}} treff for \"{{searchTerm}}\"",
+  "search_summary_plural": "{{count}} treff for \"{{searchTerm}}\"",
   "send": "Send",
   "show": "Vis",
   "start": "Start",
